Rename : type 폴더 구분
